ref: move the tests to the file of tested function
