+++ contracts/hub/router/src/execute/token.rs
@@ -337,3 +337,498 @@ pub fn _release_voucher(
         release_amount,
     ))
 }
+
+#[cfg(test)]
+mod token_tests {
+    use cosmwasm_std::{
+        testing::{message_info, mock_env},
+        Addr, Order, Uint128,
+    };
+    use euclid::{
+        chain::ChainUid,
+        cross_chain_user::CrossChainUser,
+        error::ContractError,
+        limit::Limit,
+        msgs::cross_chain_config::CrossChainConfig,
+        recipient::Recipient,
+        token::{Token, TokenType},
+    };
+
+    use crate::{
+        contract::execute,
+        state::{ESCROW_BALANCES, LOCKED_CHAINS, PENDING_RELEASE_VOUCHER, RELEASE_FEES},
+        tests::{
+            fixtures::{initialized, transfer_deps, voucher_deps},
+            tests::tests::{make_native_recipient, seed_virtual_balance, MockDeps},
+        },
+    };
+    use euclid::msgs::router::ExecuteMsg;
+    use rstest::*;
+    // -----------------------------------------------------------------------
+    // WithdrawVoucher: error cases (table-driven)
+    // -----------------------------------------------------------------------
+
+    #[rstest]
+    #[case::invalid_denom("uatom", false, ContractError::InvalidDenom {})]
+    #[case::locked_chain("uusdc", true, ContractError::new("Chain is locked"))]
+    fn test_withdraw_voucher_error_cases(
+        mut initialized: MockDeps,
+        #[case] registered_denom: &str,
+        #[case] lock_chain: bool,
+        #[case] expected_error: ContractError,
+    ) {
+        use euclid::{
+            msgs::{cross_chain_config::CrossChainConfig, router::TokenDenom},
+            token::TokenType,
+        };
+
+        use crate::{
+            state::TOKEN_DENOMS,
+            tests::tests::tests::{seed_chain1_native, seed_virtual_balance},
+        };
+
+        let creator = initialized.api.addr_make("creator");
+        let chain_uid = ChainUid::create("chain1".to_string()).unwrap();
+        let token = Token::create("usdc".to_string()).unwrap();
+
+        seed_virtual_balance(&mut initialized);
+        seed_chain1_native(&mut initialized);
+        let locked = if lock_chain {
+            vec![chain_uid.clone()]
+        } else {
+            vec![]
+        };
+        LOCKED_CHAINS
+            .save(initialized.as_mut().storage, &locked)
+            .unwrap();
+        TOKEN_DENOMS
+            .save(
+                initialized.as_mut().storage,
+                token.clone(),
+                &vec![TokenDenom {
+                    chain_uid: chain_uid.clone(),
+                    token_type: TokenType::Native {
+                        denom: registered_denom.to_string(),
+                    },
+                }],
+            )
+            .unwrap();
+
+        let res = execute(
+            initialized.as_mut(),
+            mock_env(),
+            message_info(&creator, &[]),
+            ExecuteMsg::WithdrawVoucher {
+                token,
+                amount: Uint128::new(100),
+                recipient: make_native_recipient(
+                    chain_uid,
+                    "recipient",
+                    "uusdc",
+                    Uint128::new(100),
+                ),
+                cross_chain_config: CrossChainConfig::default(),
+            },
+        );
+        assert_eq!(res.unwrap_err(), expected_error);
+    }
+
+    #[rstest]
+    fn test_withdraw_voucher_unregistered_token_fails(mut initialized: MockDeps) {
+        let creator = initialized.api.addr_make("creator");
+
+        seed_virtual_balance(&mut initialized);
+
+        let res = execute(
+            initialized.as_mut(),
+            mock_env(),
+            message_info(&creator, &[]),
+            ExecuteMsg::WithdrawVoucher {
+                token: Token::create("usdc".to_string()).unwrap(),
+                amount: Uint128::new(100),
+                recipient: make_native_recipient(
+                    ChainUid::create("chain1".to_string()).unwrap(),
+                    "recipient",
+                    "uusdc",
+                    Uint128::new(100),
+                ),
+                cross_chain_config: CrossChainConfig::default(),
+            },
+        );
+        assert!(res.is_err());
+    }
+
+    #[test]
+    fn test_withdraw_voucher_happy_path() {
+        let mut deps = voucher_deps();
+        let creator = deps.api.addr_make("creator");
+        let chain_uid = ChainUid::create("chain1".to_string()).unwrap();
+        let token = Token::create("usdc".to_string()).unwrap();
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            message_info(&creator, &[]),
+            ExecuteMsg::WithdrawVoucher {
+                token: token.clone(),
+                amount: Uint128::new(200),
+                recipient: make_native_recipient(
+                    chain_uid.clone(),
+                    "recipientaddr",
+                    "uusdc",
+                    Uint128::new(200),
+                ),
+                cross_chain_config: CrossChainConfig::default(),
+            },
+        )
+        .unwrap();
+
+        assert_eq!(
+            res.messages.len(),
+            2,
+            "expected burn + ibc release submessages"
+        );
+
+        let escrow = ESCROW_BALANCES
+            .load(
+                deps.as_ref().storage,
+                (token.to_string(), chain_uid.clone()),
+            )
+            .unwrap();
+        assert_eq!(escrow, Uint128::new(300));
+
+        let pending: Vec<_> = PENDING_RELEASE_VOUCHER
+            .range(deps.as_ref().storage, None, None, Order::Ascending)
+            .collect::<Result<_, _>>()
+            .unwrap();
+        assert_eq!(pending.len(), 1);
+        assert_eq!(pending[0].1.total_amount, Uint128::new(200));
+        assert_eq!(pending[0].1.release_fee_amount, Uint128::zero());
+
+        let released_attr = res
+            .attributes
+            .iter()
+            .find(|a| a.key == "released_amount")
+            .unwrap();
+        assert_eq!(released_attr.value, "200");
+    }
+
+    #[test]
+    fn test_withdraw_voucher_with_release_fee() {
+        let mut deps = voucher_deps();
+        let creator = deps.api.addr_make("creator");
+        let chain_uid = ChainUid::create("chain1".to_string()).unwrap();
+        let token = Token::create("usdc".to_string()).unwrap();
+
+        RELEASE_FEES
+            .save(
+                deps.as_mut().storage,
+                (token.clone(), chain_uid.clone()),
+                &Uint128::new(10),
+            )
+            .unwrap();
+
+        execute(
+            deps.as_mut(),
+            mock_env(),
+            message_info(&creator, &[]),
+            ExecuteMsg::WithdrawVoucher {
+                token: token.clone(),
+                amount: Uint128::new(200),
+                recipient: make_native_recipient(
+                    chain_uid.clone(),
+                    "recipientaddr",
+                    "uusdc",
+                    Uint128::new(200),
+                ),
+                cross_chain_config: CrossChainConfig::default(),
+            },
+        )
+        .unwrap();
+
+        // escrow reduced by release_amount_after_fee = 200 - 10 = 190
+        let escrow = ESCROW_BALANCES
+            .load(
+                deps.as_ref().storage,
+                (token.to_string(), chain_uid.clone()),
+            )
+            .unwrap();
+        assert_eq!(escrow, Uint128::new(310));
+
+        let pending: Vec<_> = PENDING_RELEASE_VOUCHER
+            .range(deps.as_ref().storage, None, None, Order::Ascending)
+            .collect::<Result<_, _>>()
+            .unwrap();
+        assert_eq!(pending[0].1.total_amount, Uint128::new(200));
+        assert_eq!(pending[0].1.release_fee_amount, Uint128::new(10));
+    }
+    // -----------------------------------------------------------------------
+    // TransferVoucher
+    // -----------------------------------------------------------------------
+
+    #[rstest]
+    fn test_transfer_voucher_unregistered_token_fails(mut initialized: MockDeps) {
+        let creator = initialized.api.addr_make("creator");
+
+        seed_virtual_balance(&mut initialized);
+
+        let res = execute(
+            initialized.as_mut(),
+            mock_env(),
+            message_info(&creator, &[]),
+            ExecuteMsg::TransferVoucher {
+                token: Token::create("usdc".to_string()).unwrap(),
+                amount: Uint128::new(100),
+                recipient: vec![euclid::recipient::Recipient {
+                    recipient: euclid::cross_chain_user::CrossChainUser::new(
+                        ChainUid::vsl_chain_uid().unwrap(),
+                        "recipient_addr".to_string(),
+                    ),
+                    amount: Limit::LessThanOrEqual(Uint128::new(100)),
+                    denom: TokenType::Voucher {},
+                    forwarding_message: None,
+                    unsafe_refund_as_voucher: None,
+                }],
+            },
+        );
+        assert!(res.is_err());
+    }
+
+    #[test]
+    fn test_transfer_voucher_to_voucher_recipient_happy_path() {
+        let mut deps = transfer_deps();
+        let sender = Addr::unchecked("sender_address");
+        let token = Token::create("usdc".to_string()).unwrap();
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            message_info(&sender, &[]),
+            ExecuteMsg::TransferVoucher {
+                token,
+                amount: Uint128::new(100),
+                recipient: vec![euclid::recipient::Recipient {
+                    recipient: euclid::cross_chain_user::CrossChainUser::new(
+                        ChainUid::vsl_chain_uid().unwrap(),
+                        "recipient_addr".to_string(),
+                    ),
+                    amount: Limit::LessThanOrEqual(Uint128::new(100)),
+                    denom: TokenType::Voucher {},
+                    forwarding_message: None,
+                    unsafe_refund_as_voucher: None,
+                }],
+            },
+        )
+        .unwrap();
+
+        assert_eq!(
+            res.messages.len(),
+            1,
+            "expected 1 virtual balance transfer submsg"
+        );
+        assert_eq!(
+            res.attributes
+                .iter()
+                .find(|a| a.key == "transferred_amount")
+                .unwrap()
+                .value,
+            "100"
+        );
+        assert_eq!(
+            res.attributes
+                .iter()
+                .find(|a| a.key == "release_initiated")
+                .unwrap()
+                .value,
+            "0"
+        );
+    }
+
+    #[test]
+    fn test_transfer_voucher_self_transfer_no_submsg() {
+        let mut deps = transfer_deps();
+        let sender = Addr::unchecked("senderaddr");
+        let token = Token::create("usdc".to_string()).unwrap();
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            message_info(&sender, &[]),
+            ExecuteMsg::TransferVoucher {
+                token,
+                amount: Uint128::new(100),
+                recipient: vec![euclid::recipient::Recipient {
+                    recipient: euclid::cross_chain_user::CrossChainUser::new(
+                        ChainUid::vsl_chain_uid().unwrap(),
+                        sender.to_string(),
+                    ),
+                    amount: Limit::LessThanOrEqual(Uint128::new(100)),
+                    denom: TokenType::Voucher {},
+                    forwarding_message: None,
+                    unsafe_refund_as_voucher: None,
+                }],
+            },
+        )
+        .unwrap();
+
+        assert!(
+            res.messages.is_empty(),
+            "expected no submessages for self-transfer"
+        );
+        assert_eq!(
+            res.attributes
+                .iter()
+                .find(|a| a.key == "transferred_amount")
+                .unwrap()
+                .value,
+            "100"
+        );
+    }
+
+    // -----------------------------------------------------------------------
+    // WithdrawVoucher: escrow-cap and Limit variants
+    // -----------------------------------------------------------------------
+
+    /// When the requested amount exceeds the escrow balance the release is
+    /// silently capped at the available escrow.
+    #[test]
+    fn test_withdraw_voucher_capped_at_escrow_balance() {
+        let mut deps = voucher_deps(); // escrow = 500
+        let creator = deps.api.addr_make("creator");
+        let chain_uid = ChainUid::create("chain1".to_string()).unwrap();
+        let token = Token::create("usdc".to_string()).unwrap();
+
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            message_info(&creator, &[]),
+            ExecuteMsg::WithdrawVoucher {
+                token: token.clone(),
+                amount: Uint128::new(1_000),
+                recipient: make_native_recipient(
+                    chain_uid.clone(),
+                    "recipientaddr",
+                    "uusdc",
+                    Uint128::new(1_000),
+                ),
+                cross_chain_config: CrossChainConfig::default(),
+            },
+        )
+        .unwrap();
+
+        // Escrow fully drained (not negative).
+        let escrow = ESCROW_BALANCES
+            .load(deps.as_ref().storage, (token.to_string(), chain_uid))
+            .unwrap();
+        assert_eq!(escrow, Uint128::zero());
+
+        // PENDING_RELEASE records the actual released amount (500), not the requested 1000.
+        let pending: Vec<_> = PENDING_RELEASE_VOUCHER
+            .range(deps.as_ref().storage, None, None, Order::Ascending)
+            .collect::<Result<_, _>>()
+            .unwrap();
+        assert_eq!(pending[0].1.total_amount, Uint128::new(500));
+
+        // released_amount attribute also reflects the cap.
+        assert_eq!(
+            res.attributes
+                .iter()
+                .find(|a| a.key == "released_amount")
+                .unwrap()
+                .value,
+            "500"
+        );
+    }
+
+    /// GreaterThanOrEqual limit fails when the escrow can't satisfy the minimum.
+    #[test]
+    fn test_withdraw_voucher_gte_limit_fails_when_escrow_too_low() {
+        let mut deps = voucher_deps(); // escrow = 500
+        let creator = deps.api.addr_make("creator");
+        let chain_uid = ChainUid::create("chain1".to_string()).unwrap();
+        let token = Token::create("usdc".to_string()).unwrap();
+
+        // Request 200 but require at least 600 (escrow=500 < 600).
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            message_info(&creator, &[]),
+            ExecuteMsg::WithdrawVoucher {
+                token,
+                amount: Uint128::new(200),
+                recipient: Recipient {
+                    recipient: CrossChainUser::new(chain_uid, "recipientaddr".to_string()),
+                    amount: Limit::GreaterThanOrEqual(Uint128::new(600)),
+                    denom: TokenType::Native {
+                        denom: "uusdc".to_string(),
+                    },
+                    forwarding_message: None,
+                    unsafe_refund_as_voucher: None,
+                },
+                cross_chain_config: CrossChainConfig::default(),
+            },
+        );
+        assert!(matches!(
+            res.unwrap_err(),
+            ContractError::InsufficientAmount { .. }
+        ));
+    }
+
+    // -----------------------------------------------------------------------
+    // TransferVoucher: multi-recipient sequential allocation
+    // -----------------------------------------------------------------------
+
+    /// When multiple recipients are listed, amount flows through them in order;
+    /// each recipient absorbs up to its limit and the remainder goes to the next.
+    #[test]
+    fn test_transfer_voucher_splits_amount_across_recipients() {
+        let mut deps = transfer_deps();
+        let sender = Addr::unchecked("sender_address");
+        let token = Token::create("usdc".to_string()).unwrap();
+        let vsl_chain = ChainUid::vsl_chain_uid().unwrap();
+
+        // recipient1 takes up to 60, recipient2 takes the remainder (40).
+        let res = execute(
+            deps.as_mut(),
+            mock_env(),
+            message_info(&sender, &[]),
+            ExecuteMsg::TransferVoucher {
+                token,
+                amount: Uint128::new(100),
+                recipient: vec![
+                    Recipient {
+                        recipient: CrossChainUser::new(vsl_chain.clone(), "addr_one".to_string()),
+                        amount: Limit::LessThanOrEqual(Uint128::new(60)),
+                        denom: TokenType::Voucher {},
+                        forwarding_message: None,
+                        unsafe_refund_as_voucher: None,
+                    },
+                    Recipient {
+                        recipient: CrossChainUser::new(vsl_chain, "addr_two".to_string()),
+                        amount: Limit::LessThanOrEqual(Uint128::new(100)),
+                        denom: TokenType::Voucher {},
+                        forwarding_message: None,
+                        unsafe_refund_as_voucher: None,
+                    },
+                ],
+            },
+        )
+        .unwrap();
+
+        // Two separate virtual-balance transfer submessages.
+        assert_eq!(
+            res.messages.len(),
+            2,
+            "expected one submsg per non-self recipient"
+        );
+        // Amounts reflected in attributes.
+        let attrs: std::collections::HashMap<_, _> = res
+            .attributes
+            .iter()
+            .map(|a| (a.key.as_str(), a.value.as_str()))
+            .collect();
+        assert_eq!(attrs["transfer_id_0_amount"], "60");
+        assert_eq!(attrs["transfer_id_1_amount"], "40");
+        assert_eq!(attrs["transferred_amount"], "100");
+    }
+}
