+++ contracts/hub/router/src/ibc/receive/base.rs
@@ -127,3 +127,77 @@ pub fn reusable_internal_call(
 
     Ok(response)
 }
+
+#[cfg(test)]
+mod tests {
+
+    use euclid::{
+        chain::ChainUid,
+        cross_chain_user::CrossChainUser,
+        error::ContractError,
+        token::{Token, TokenType, TokenWithDenom},
+    };
+    use euclid_ibc::router_ibc::RouterCrossChainExecuteMsg;
+
+    use crate::{
+        state::{LOCKED_CHAINS, STATE},
+        tests::{
+            fixtures::initialized,
+            tests::tests::{call_reusable, register_denom_msg, MockDeps},
+        },
+    };
+
+    use rstest::*;
+
+    #[rstest]
+    fn test_receive_dispatch_contract_locked(mut initialized: MockDeps) {
+        let mut state = STATE.load(initialized.as_ref().storage).unwrap();
+        state.locked = true;
+        STATE.save(initialized.as_mut().storage, &state).unwrap();
+
+        let chain_uid = ChainUid::create("chain1".to_string()).unwrap();
+        let result = call_reusable(
+            &mut initialized,
+            register_denom_msg(chain_uid.clone()),
+            chain_uid,
+        );
+        assert_eq!(result.unwrap_err(), ContractError::ContractLocked {});
+    }
+
+    #[rstest]
+    fn test_receive_dispatch_chain_locked(mut initialized: MockDeps) {
+        let chain_uid = ChainUid::create("chain1".to_string()).unwrap();
+        LOCKED_CHAINS
+            .save(initialized.as_mut().storage, &vec![chain_uid.clone()])
+            .unwrap();
+
+        let result = call_reusable(
+            &mut initialized,
+            register_denom_msg(chain_uid.clone()),
+            chain_uid,
+        );
+        assert_eq!(result.unwrap_err(), ContractError::DeregisteredChain {});
+    }
+
+    #[rstest]
+    fn test_receive_dispatch_chain_uid_mismatch(mut initialized: MockDeps) {
+        let chain1 = ChainUid::create("chain1".to_string()).unwrap();
+        let chain2 = ChainUid::create("chain2".to_string()).unwrap();
+        // sender is from chain2 but chain_uid param is chain1
+        let msg = RouterCrossChainExecuteMsg::RegisterDenom {
+            sender: CrossChainUser::new(chain2, "user".to_string()),
+            token: TokenWithDenom {
+                token: Token::create("usdc".to_string()).unwrap(),
+                token_type: TokenType::Native {
+                    denom: "uusdc".to_string(),
+                },
+            },
+            tx_id: "tx1".to_string(),
+        };
+        let result = call_reusable(&mut initialized, msg, chain1);
+        assert_eq!(
+            result.unwrap_err(),
+            ContractError::new("Chain UID mismatch")
+        );
+    }
+}
